requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
boto3==1.12.36
botocore==1.15.36
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crontab==0.7.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-js-asset==1.2.2
django-nested-inline==0.4.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-utils==0.0.2
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
hashids==1.2.0
hiredis==1.0.1
hyperlink==19.0.0
iamport-rest-client==0.8.1
idna==2.9
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.19.4
numpy-stl==2.12.0
openapi-codec==1.3.2
packaging==20.3
Pillow==7.0.0
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-utils==2.4.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.8
zope.interface==5.1.0
